Implementation of AvroSedes.serialize (#68)
* Update SchemaConverters.scala

If the reader returns Object instead of GenericRecord the deserialize method will support of the avro schemas.
It currently casts to GenericRecord and it doesn't work for simple type and union schemas.

* Update README.md

Created new block for support of Avro schemas with relevant examples.

* Implementing AvroSedes.serialize

Implemented the method in a way that it serialize the data to byte arrays using their Avro schema.
It supports all the possible Avro schemas and types.

* Update SchemaConverters.scala

Fixed the formatting in method AvroSedes.serialize

* Avro serialization tests

Additional tests for serialization/deserialization of different Avro schemas

* Update AvroRecordSuite.scala

Formatting the new tests

* Update README.md

Modified as per code review

* Update AvroRecordSuite.scala

Extended the test to cover all the primitive and union schemas supported by Avro.

* Update AvroRecordSuite.scala

Added missing import of java.nio.ByteBuffer

* Update README.md

Formating in the Others chapter.