In secure cluster put hbase-site in SPARK_CONF_DIR
Running a Spark job in a HDP cluster secured by Kerberos and passing the hbase-site.xml with "--files" or in the classpath never worked for me.
As described in https://community.hortonworks.com/content/supportkb/48988/how-to-run-spark-job-to-interact-with-secured-hbas.html (point 2), the only solution which worked was to include the hbase-site.xml directly in the Spark conf directory of our Edge node (/etc/spark/conf).
I left the option to pass hbase-site.xml with "--files" as I am not certain that it does not work in other environments.