Delegate to specific coder when scanning for Strings, Array[Byte], etc (#116)
## What changes were proposed in this pull request?

The calculation of the range for the scanner is not correct for types String, Array[Byte], Byte, UTF8String. The serialized representations should be obtained from the coder. In the case of PrimitiveType and Phoenix, the serializations happen to match the code in ScanRange.scala. However, I have a custom coder that does not have the same serialization.

## How was this patch tested?

All existing unit tests still pass